create newModel generator in pathHelp so we minimize duplicated code in routes
