+++ ui/app/utils/openapi-to-attrs.js
@@ -55,6 +55,7 @@ export const combineAttributes = function(oldAttrs, newProps) {
       newFields.push(prop);
     }
   }
+  debugger; //eslint-disable-line
   return { attrs: newAttrs, newFields };
 };
